Move movements sync-schema to separate file
